refactor: rename module nuvlabox to nuvlaedge_docker
